Major refactor. Game no longer works.
But, tests! Game objects now respond to a defined set of
GameActions, found in message.py. After processing an action,
the Game.handle() function returns and the state is unchanged
until another action is passed in.

Old methods are still in the file, prefixed with '_'.

A re-implementation of the client can be done where the
game actions can be generated and passed in.

Also added __init__.py files, so if your PYTHONPATH is correct,
you can "import cloaca.gtr". This is used for the tests.