Add game log that records what's happening.
It's a member of the GameState object and gets posted
to the curses_gui. I had to rework the order of a lot of the
handle_*() methods so that nothing gets logged if there's an error.

This required lots more refactoring for the cases where a helper method
was used, specifically move_legionary_cards(), and construct().

There are more cases where a log message should be added. Also, I think
I broke legionary and GIVECARDS. That will be fixed next.